sensing: loop sensors in runtime module
add runtime module, implement loop sensors in runtime module
