+++ subsys/sensing/CMakeLists.txt
@@ -5,6 +5,7 @@ zephyr_library_include_directories(include)
 
 zephyr_library_sources(
 	sensor_mgmt.c
+	runtime.c
 	sensing.c
 	sensing_sensor.c
 )
