+++ subsys/sensing/runtime.c
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2022 Intel Corporation.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+#include <zephyr/logging/log.h>
+#include <zephyr/sensing/sensing_sensor.h>
+#include "sensor_mgmt.h"
+
+LOG_MODULE_DECLARE(sensing, CONFIG_SENSING_LOG_LEVEL);
+
+int loop_sensors(struct sensing_context *ctx)
+{
+	return 0;
+}
